postgres: add script for automatic resyncing
Currently resyncs of replicas are done by hand, more or less chaining
the commands in this script. There is a cookbook for initialising
postgres hosts that is currently broken and uses a number of similar
chained commands. Simplify this logic by having the cookbook run a
single command - this one.

Bug: T275381
Change-Id: Ib4815ea6fc4745cc93cfbc905a25df7f2f0f7e9e