Remove unused Kerberos settings from sshd config file
KerberosAuthentication, KerberosGetAFSToken, KerberosOrLocalPasswd and
KerberosTicketCleanup are only relevant for the default Kerberos
implementation in OpenSSH. It is rather limited and instead most
distros use a patch which adds support via GSSAPI, but which got
rejected in https://bugzilla.mindrot.org/show_bug.cgi?id=1242.

Our relevant config settings are GSSAPIAuthentication and
GSSAPICleanupCredentials, so drop the Kerberos ones to avoid
confusion.

Change-Id: Ia3ecd172d8af75683b5ec93e32c7870355954b06