Set a symlink for the host keytab
OpenSSH searches for the host keytab in /etc/keytab. We deploy all keytabs
centrally via Puppet in /etc/security/keytabs, so add a symlink as a fallback.

In theory we could also pass KRB5_KTNAME via the openssh systemd unit, but
it still seems like a sensible default for other applications as well.

Bug: T244840
Change-Id: I175b8e20fe103a7605e18653d768c811975985c4