admin: update admin modules so users without sshkeys get nologin shell
We have a number of users which require a unix account but don't actually
need to logon and obtain an tty.  As such update the admin module to
default users with no sshkey to have a shell of nologin.

W also set the shell of absent users to nologin

Hosts: an-master1001.eqiad.wmnet,cumin1001.eqiad.wmnet
Change-Id: I8ba398fe2c5ea355923be5e6e49d35cfa4d74921