openstack-glance: enable copy-on-write inside ceph
In order for the images in ceph to properly work as COW without a local
copy being downloaded to the disk and then uploaded again to RBD (which
is really just relying on it being sparse, not COW), we need to allow
nova-compute to see the location of the image in ceph and read it.

That way, all operations happen on ceph instead of the weird flip flop
that slows everything down and uses more disk space.

Bug: T275430
Change-Id: Ib6beee9d5c8d4fcdad209f217d6ed6c807fa8911