labs_lvm: change the default disk in case we ever create an odd image
This module only creates a volume group and uses the $disk param if
it is missing the volume group "vd". This is there on our images by
default, but going forward, the disk name has changed.

So this should be a noop unless someone creates a new VM that somehow
is missing the volume group.

Bug: T275585
Change-Id: I9bdb2ccbf0a4a05f6a7367a2041dfca8c77561e5