role::druid::public::worker: tune cache settings
We are now running a 5+ cluster, and I suspect that the big number of
segments per dataset that we keep makes it behave more towards what
upstreams calls "big production cluster". While doing a roll restart
this morning of the Druid public cluster, I got AQS alerts right after
the first historical stop/start, and I had to force a complete roll
restart of the brokers to restore the service (they were locked up
exactly how it happens when we drop a datasource). I kept going with the
roll restarts as test, and I didn't get any problem. One big difference
was the fact that after the broker roll restart the local cache was
gone, so I have the feeling that the brokers suffer a lot when the cache
is full (since it may cause LRU/evictions/etc..).

https://druid.apache.org/docs/latest/querying/caching.html#query-caching-on-brokers

Bug: T270173
Change-Id: I1998038a73fa7249aac55f2bbe735c9f3c5f8ae4