Don't drop namespace if it's already gone
In some cases (like manual reloads), old entries in alias map
for wdqs namespaces (mostly categories) may be incorrect, causing
replaceNamespace function to fail on deleting a namespace (and
in consequence the script if it has set -e). This change checks
if namespace exists and doesn't try to delete it when it doesn't.

Bug: T269331
Change-Id: Ib4cdb4f98e80766bf14e3a1deb51f80aa8868497