kubelet: Support the IPv6 feature gate
Why:
IPv6 is an alpha feature on 1.16 and needs a variety of tunables
(https://kubernetes.io/docs/concepts/services-networking/dual-stack/) to
fully work. However we don't need IPv6 to fully work, we want to
retain the ability to have routeable IPv6 address assigned to pods. This
used to happen via calico with kubelet being totally unaware of the
fact, however on 1.16+ we need to inform the kubelet as well.

We are on purpose here NOT adding support for IPv6 feature gate
enablement on kube-apiserver, kube-controller-manager and kube-proxy as
the feature is alpha grade and not deemed stable yet on our version.

What:
Add a hiera populate parameter named $kubelet_ipv6 that enable IPv6 on
the kubelet. Default it to false and populate accordingly the hieradata/
files to enable it just on staging-codfw

This has been tested manually on a staging-codfw node and works fine in
my limited tests

Change-Id: I54a042731f60dc02494907022cb8115fae052c50