mediawiki::alerts: add per cluster error/fatals rate alert
What?
This patch creates the log_mediawiki_servergroup_level_channel_doc_count
aggregation which includes the "servergroup" label.

Why?
When the "MediaWiki exceptions and fatals per minute" fired,
we could't tell which cluster was the one producing the
errors, unless we went to logstash.

Notes: After merging this patch, we need to update the
graphs in
https://grafana.wikimedia.org/d/000000438/mediawiki-alerts

Bug: T262078
Change-Id: I2cf3a97baec338216a40d56ec5318a48654a2791