wmcs-wikireplica-dns.py: create a new session for each domain
We are seeing intermittent failures in this script.  It seems to happen
after long runs so I'm concerned that it's a timeout someplace.

I have no real reason to think that this patch will fix things, but
if our sessions are expiring this should work around that.

Note that sessions don't persist on the server side (thanks to fernet tokens)
so this should be basically harmless other than slowing things down a bit
on the client.

Bug: T274956
Change-Id: I5b4b75c019326a155e7261bfbcf052e283a4707a