phabricator: replace task dump cron with timer and switch to weekly
We are replacing the cronjob with a systemd timer.

The script itself has been modified to drop stdout to /dev/null
so that part isn't needed anymore in shell.

Also we noticed the dump might take longer than 24 hours and
might not be used by anyone, so at the same time switching
it from daily to weekly for now.

Bug: T273673
Change-Id: Iaaab0b13192aa20e75a88780d925f9c52d279110