typos: remove 'bullsey'
'bullsey' is a substring of the correct spelling 'bullseye',
so it triggers the typo check if you add the correct version.

So I got jerkins -1 on unrelated changes to a file like in
I998eaf3084c6844da7.

This was added in cebfc7f133b2ea10bc85f1. Removing that one
variant.

Change-Id: I2cdc69dceaa9fa306deee81d892e77c4358ca1a2