wmcs.postrgesql.osm_primary: add missing required engine param
The parameter was added here:
  https://gerrit.wikimedia.org/r/c/operations/puppet/+/644482

But it did not fail the tests because there were none xd
So this also adds a dummy test for that module.

Note that I had to add a new case for postgres/common.pp to support
buster distros in order to run the tests.

Bug: T276039
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Ib6820f5d01f46572f2f5a8480b89c022e130e53f