Configure spark to work better with conda environments
This patch does the following:

1. When running pyspark2 CLI, defaults to using ipython for PYSPARK_DRIVER_PYTHON
2. When a conda enviroment is active and master=yarn, defaults PYSPARK_PYTHON to anaconda-wmf python

1. is inconsequential, but it does affect pyspark2 even if a conda env is not active.
This should just be an improvement in the default python CLI when using the pyspark REPL.

2. Should be a no-op unless conda is active. If a conda env is active and spark is
being launched in YARN, we need to make sure that the python that is used on
remote workers actually exists.  We install anaconda-wmf everywhere (in
/usr/lib/anaconda-wmf), so defaulting to this for PYSPARK_PYTHON
is safe and the right thing to do.  The user can still override PYSPARK_PYTHON
as needed, e.g. to ship a packed conda or virtualenv.

We also set ARROW_PRE_0_15_IPC_FORMAT to 1, to work around
https://issues.apache.org/jira/browse/SPARK-29367 until
we upgrade Spark > 2.4.4.

Bug: T272313
Change-Id: I6be939511081518390614cde9f29a11414434b4a