profile::templates::services_proxy: switch to ::1 when listen_ipv6 is true
After discussing in T255568, we determined that service proxy might
not even need to listen to 127.0.0.1, let alone 0.0.0.0. In order
to switch service proxy to ipv6, and avoid any risks related to
ipv4_compat, we can simply have it listen to ::1, when listen_ipv6
is enabled and remove ipv4_compat.

Bug: T255568
Change-Id: I45143f83c458749d7eb8246b619ebfdbb8b53a64