Rsync private mediawiki files to releases server
Allows an rsync of the /srv/mediawiki-staging/private directory from the
deployment server to the releases server.

Private mediawiki files are needed on the production docker image for
mediawiki.

Adds the releases::mediawiki::private profile so that the rsync module
can be used to copy the /srv/mediawiki-staging/private folder from the
deployment server to the releases server.

Bug: T276145
Change-Id: I109310e0c0e88d3d85bb1d1f13f985b0c69b3998