P:profile::client::httpd: add param override the attribute delimiter
We default this to ',' as this is the upstream default.  however for use
it makes much more sense to use anything but ',' as the important
attribute to use is the memberof attribute which is a ldapdn and this
contains commers naturarly.

However some set ups already use this mapping (librenms at the very
least) we would need to investigate to see there are any other instances
and what other local modiufications may be required

Hosts: cumin:R:profile::idp::client::httpd::site
Change-Id: I63031fc296e907a388f9e5c804d72a50e9258e39