hadoop: bump HDFS Namenode heap size
I noticed high GC pauses in the Namenode logs, identified by the
org.apache.hadoop.util.JvmPauseMonitor thread. This caused some lag in
answering to the HDFS Zkfc daemon on an-master1001, that in turn forced
a failover to an-master1002 for safety.
As indicated by the following guide (that we have followed in the past),
we should now be using a higher heap size, since we are currently
storing ~54M files.

https://docs.cloudera.com/HDPDocuments/HDP2/HDP-2.6.3/bk_command-line-installation/content/configuring-namenode-heap-size.html

Performance note: the JVM stops using compressed object pointers after
the 32G heap mark, switching to 64 bits and hence allocating more space
for object references. There is no clear indication about what's best to
do after the 32G, but in general there seems to be some consensus in
bumping to 36/38 directly, that is in line with the previous link.

Change-Id: I0c84bbfe44352e748137cb5beb9ef2d34b12c165