parsercache.my.cnf: innodb_change_buffering = none
Following some mariadb tickets and discussions, it might be beneficial
to set innodb_change_buffering = none and see if the crashes do not
repeat.
Note: this do not prevent possible corrupted data to be fixed, it might
just prevent future occurrences.

I haven't seen any performance issues so far with it.
Making it permanent on the config

Bug: T263443
Change-Id: I37cb42eced81b11b2146aca431737617a2da5fe3