nova.conf: Adjust live migration timing values
All of these changes are intended to make live migrations more
likely to succeed and to complete sooner. The tradeoff is that
migrated VMs are more likely to hesitate (for up to 1 second)
during migration.

The previous default values meant that we would wait 12 minutes
before allowing a VM to pause for half a second.  The new settings
mean that we will allow a VM to pause for one second after four
minutes of trying.

This patch also allows up to three concurrent migrations (replacing
a previous default of 1).  This should allow host draining to proceed
if a VM or two get stuck in a migration state for whatever reason.

The whole story of this patch is: "make VM reliability worse
in order to make cloud admins' lives easier". I'm not 100%
sure this is the right trade-off, but I'm also not sure that
the amount this degrades performance will actually be noticeable
to anyone.

Bug: T276344
Change-Id: Ica7818aea49885cf5417fc1d7586ee3a6129db7e