wikireplicas: expose actor_user = NULL (IPs) again in actor view
In our haste to ensure the filtering correctly removed values from the
actor table in the wikireplicas, we removed IP users as well. This is
not the intent and breaks some tools and limits valuable information about
edits and vandalism. Don't merge this until I've tested it to make sure
it retains the correct sense of filtering while restoring the anonymous
IP-based actors to view.

Bug: T276698
Bug: T276628
Bug: T276124
Change-Id: Ie7b66bff3348f84f032a984bf26cda6a764c626a