admin: create new admin group mailman3-roots
A new admin group for root-level access on mailman3
VMs.

To be applied temporarily to a new VM but not
to existing mailman2 production machines.

This adds the new group and adds it to the
lists3 role that already exists but is not
applied to a VM yet.

Adding members to it will be separate.

Bug: T276712
Change-Id: I8569ee9f3fffffb8475537349f149bf9768644fc