aqs: make aqs1010 a separate AQS cluster
To begin the process of migrating data to a new buster-based AQS
cluster in advance of upgrading to cassandra 3.11, make a new cluster
(based on seed definitions and the fact that the hosts will be
firewalled off from each other by default). For now, add a single node
for testing purposes.

Bug: T257572
Change-Id: Ia0e965f01b710b27207780e2e77d1afb1f32047a