Update cache config for superset 1.0
Superset has 2 types of caching now: DATA_CACHE_CONFIG for results and
CACHE_CONFIG for metadata. This configures both, with different key
prefixes.

Currently caching is disabled in production, so this change will only
affect staging, where we are testing Superset 1.0 to see if this fixes
the caching permission issue we found.

Hosts: an-tool1005.eqiad.wmnet,an-tool1010.eqiad.wmnet
Bug: T273850
Change-Id: If1bc1ff376584ee6ebbd9db8174c5c68b1c4b8d1