gitlab: open firewall for 80,443. use drange to limit to service IP
Since we now have a second IP to run public services, https and
ssh, on, we should open the firewall on ports 80 and 443
but only for traffic to the service IP, not for the main
server IP.

So we are using drange for that which isn't used a lot
but makes sense in this situation.

That way SSH to the the server IP gitlab1001 is still limited
to come from bastion hosts.

per T276144#6890920 we can wait or merge now

Bug: T276144
Change-Id: I3f4159d6a3f6aa415d21a7bfaab54da885eff103