cinderutils::ensure: Refactor and add a 'mount' resource
This change is mostly in service of getting puppet to manage
the mount that is created by our call to prepare_cinder_volume.

The custom fact now yields all (presumed) cinder volumes so we
can look up the uuid of a volume after it's mounted; the job
of detecting 'newness' is left to the puppet class which
just checks to see w/not the volume has an assigned mountpoint.

Bug: T269511
Change-Id: Ic8dcdf9b830f5606123d936b4c0e9f534c08989a