netbase: add new module to manage /etc/services
This Cr adds a new module intended to provide an interface in to service
definitions. In the basic form it exposes all the service
normally found in /etc/services and provides a few helper functions
to allow users to get information about a specific service(s) in puppet
e.g.
 netbases::service('ssh') >> {
   'protocols' => ['tcp'],
   'port' => 22,
   'description' => 'SSH Remote Login Protocol'}
 }
 netbases::port(22) >> {
   'protocols' => ['tcp'],
   'port' => 22,
   'description' => 'SSH Remote Login Protocol'}
 }

We are also able to inject our own service ports into the module
which allows us to also do something like:
  netbase::service('citoid') >> {
   'protocols' => ['tcp'],
   'port' => 4003,
  }

Further there is the option to also manage the /etc/services file which
means our custom services will show up in command like ss/netstat etc.

The data used to build the service list is generated using a rake task
which parses:
  https://salsa.debian.org/md/netbase/-/raw/master/etc/services
The intention of the module is that this list of ports will always be
authoritative and present in both the structures and /etc/services file
regardless of what additional services are supplied.  i.e.
if one passes
  services => {'dns' => {'protocols' => ['udp'], 'port' => 53}}
then it will be ignored in favour of the default `domain` service

The `append_aliases` option alterers this behaviour slightly in that the
service name 'dns' would get added to the aliases list.  this would
result in the following /etc/services entry

  domain 53/tcp dns
  domain 53/udp dns

instead of simply

  domain 53/tcp
  domain 53/udp

Bug: T277146
Change-Id: I841d1df43699f69f80f96b841c32616653dca257