aqs: test cassandra 3.11 on aqs1010
As a test, run cassandra 3.11 on aqs1010. Given that we're planning on
using sstableloader to import data rather than migrate the actual
datastores used by the other nodes, if the import works then we can be
reasonably sure that this migration path will work. However, if it
fails then we will have to try importing packages for and then using a
later verison of Cassandra (2.2.19 or similar).

Bug: T274119
Change-Id: I7c088301df4641f1a591e0f1d35658aab08112d6