cinderutils::ensure: Gracefully handle lvm legacy cases
As we transition from lvm-based mounts to cinder-based mounts,
it'll be nice to not have to keep track of which VMs use which
when assigning puppet roles.

This patch allows us to replace labs_lvm:volume with
cinderutils::ensure for VMs that are already provisioned
with LVM.  The cinder class will recognize that the
mounts are already present and continue to manage
them as before without worrying about whether or not
they're actually cinder-based.

Bug: T272114
Change-Id: I6cfdf1d7f6fcb2aab5f9b63d1b7d0e4b15c03a08