paws: add a hiera-controlled ip blocklist
This should provide a fairly simple mechanism for PAWS admins to block
an ip address or an ip address range (format like 10.0.0.0/16) from using
the PAWS hub (but not PAWS public, which is read-only anyway).

A doc is forthcoming, but the basic idea will be to modify the
profile::wmcs::paws::ip_blocks key in puppet hiera, adding IPs or ranges
to an array as strings. This will trigger a reload of haproxy and take
effect right away after a puppet run.

Bug: T276615
Change-Id: If954b2a070941ad4ca0a4e1ea0cfdead1dca6b0e