DHCP: remove mw2243 through mw2250
These have in common that they have
purchase date 2016-06-02  and procurement
task T134272.

They are a few days younger and separate procurement
from mw2215 through mw2243 but also 2016.

mw2251 and up are 2016-12-12 and not considered "old" yet
as opposed to the ones from June and before.

This would match what Papaul said (up to mw2250).

Change-Id: I1cbda7765972188ee6939ecb66ace62bf1322b72