gerrit: make voting scores a bit more accessible
On a change, the Gerrit scoreboard has colored chips for each of the
label scores (-2, -1, 0, +1, +2):

 ✔️   Verified      ( +2 )  jenkins-bot
 ⌛  Code-Review   ( -1 )  Jane

However they might be hard to read for some, notably colorblindness
would make it a challenge to differentiate between -1 and +1.

Make the font bold and 10% larger, that would make the minus or plus
signs stand up a bit more.

To reproduce the colorblindness, I have locally adjusted the voting
chips color background to gray.  The bold weight and slightly larger
font make -1/+1 easier to differentiate.

Bug: T256615
Change-Id: I9dbfb2af140dda4a5ee9559118ac5298525fd6a3