lvs: Set depool_threshold to .8 for upload & text
A depool threshold of 0.80 allows for two nodes out of 6 or 8 to be depooled
per cluster. Pybal checks if a server can be depooled with the following
expression:
upServerCount >= totalServerCount * depoolThreshold

upServerCount contains the number of healthy AND administratively enabled
servers.
totalServerCount contains the number of servers (healthy or unhealthy,
enabled or disabled)

Setting depoolThreshold to 0.8 allows administratively depooling up to
two servers assuming that the whole cluster is healthy.

If more than two servers of the same cluster are failing to meet pybal's
healthchecks only the first two will be depooled, the rest still will
get user traffic.

Bug: T274888
Change-Id: If24cb8817d768c8fbc17c5bb1625f23c8b60512c