idp: Make the memcached transcoder configurable
There's a regression in 6.2 in the Kryo transcoder we're currently using
and using the Java-internal one might reduce complexity in the long run
(at the cost of increased memory footprint, which isn't a huge concern
at our installation size). Make the transcoder configurable to test this
in idp-test.

Bug: T271684
Hosts: C:profile::idp
Change-Id: I3553263e40a9dc7ffdc91c3ae6a5c5e0242565e9