paws: ip blocklist needs to tcp reject if in tcp mode or change mode
This is a correction of a mistake in If954b2a070941ad4ca0
The http deny will be ignored if in tcp mode. Additionally, if I don't
switch to http mode, I cannot read host headers and safelist the public
pages. I think this is the way to go.

If I switch back to tcp mode, then we'd not allow public either, and the
deny line should be `tcp-request connection reject if is_blocked_ip`

Bug: T276615
Change-Id: I72787f6349401e9edabb0688a11506bcc7ede9d3