Revert "gerrit: make voting scores a bit more accessible"
GerritSite.css was for the GWT interface / Gerrit 2.15. Nowadays
it is only used for the login page.

Instead we should PolyGerrit wm-common-style module defined in:
./files/homedir/review_site/static/gerrit-theme.html

But since Gerrit 3.1, we can't override arbitrary css and can only
use css variables.

In short: I had to test it, did not and it doesn't work ;)

This reverts commit a9e2e25c1dcbfb00d5c52c9f21217d01bdb4fef9.

Bug: T256615
Change-Id: I857fc7813fc5def228c5eb1d5b4e73d923adea32