Add cinderutils::swap
This will look for an appropriate block device and convert
it to a swap volume.

This is in 'cinderutils' and will indeed work with a cinder volume.
The idea of using a cinder volume for swap is silly, though -- I
expect this to mostly be used with VMs build from flavors that
have a ready-made (ephemeral) swap partition.

It should work to apply this class to VMs with lvm-created
swap partitions since the first thing this class does is
check if there's already a block device formatted for swap.

Bug: T272114
Change-Id: Ie1ba35c8c259bfbfed3c9fa18c03b4d905f946c9