aqs: move import of ::passwords::aqs
Currently we require this class in the role for AQS. Technically this
is a style violation - in the process of introducing aqs_next a
similar change was blocked by tests. Instead of requiring the class
here, use hieradata lookups and the private repo to remove the class.

This review depends upon https://gerrit.wikimedia.org/r/c/labs/private/+/672441

Bug: T257572
Change-Id: Idefd9a3c30b5c9940ba44381dc65f3f50f7c4dc8