logstash: remove type setting on dlq input
The type field is not a valid ECS field and would not be present on dead
letters of ECS events.  Adding the type field here at the input triggers
the drop condition in dlq_transformer since the type field is not
present in the original event.

Remove the implicit typing on the dead_letter_queue input and rely on
the dlq_transformer to set this field in the event a dead letter of a
dead letter is generated.

Bug: T277080
Change-Id: I47f50cffdc4c1c5f8370f01cf16f95f363419c47