Nova vendordata first boot: try to work around a puppet race
The puppet module in cloud-init is handy for config but it also
starts the puppet agent. That means that a puppet run might be
already in progress when our first boot script tries to start
puppet, which causes things to happen out of order or not at all.

Stopping the agent explicitly before we do our other first boot
things should avoid that.

Change-Id: I72dd81247d6802d605461121514bf3bd280745a5