systemd::service: Add ability to monitor services
This CR adds the ability to configure an NRPE check for specific
services.  This is of disabled by default as we have the generic
`check_systemd_state` check which checks all systemd state.

The reason one might want to create a specific service check could be
to:
 * supply a specific notes_url for the service
 * supply a different admin group to be alerts when the service fails

Hosts: cumin1001.eqiad.wmnet,cp1079.eqiad.wmnet
Change-Id: I25ffde8e0246d3784d51ee44c2d81cde5df367d3