maintain-dbusers: polish things up a bit
It was widely agreed that nested ternary expressions in Python are Not
Good, despite them being fun. Here we are replacing those with something
possibly more elegant than if-elif-else, fixing a bug in the delete for
PAWS accounts and adding type hints where they aren't buggy and where
they would have really helped on the first two passes at adding PAWS to
this script.

Bug: T276284
Change-Id: If7fde880b9e24f294624092921ea3be99c28d9ac