mailman3: Use acme-chief, unify Apache configuration
We need to support two setups, using acme-chief in production and
certbot in Cloud. For both, the only difference in the Apache
configuration is the domain name (already a parameter) and where the
certificates live.

A new hiera key, "acme_chief_cert" controls which acme-chief cert to
use. If not set, we'll fallback to the default certbot paths instead.

All HTTP requests will redirect to HTTPS, though we might exempt
hyperkitty from this in the future.

Bug: T256536
Change-Id: I18647707943b10a97a3c63a28fa8715cbe29b358