Reduce buffer pool memory for dbstore1004's mariadb instances
This is an attempt to see if dbstore1004's memory usage can get more in
control over time. The dbstores currently have the following buffer pool memory
settings (for mariadb instances):
- dbstore1003: 390G
- dbstore1004: 375G
- dbstore1005: 385G

So overall dbstore1004 is inline with the others, but for some reason it
keeps ending up in using more memory.

Bug: T270112
Change-Id: I59654044c04f8327882e299aa795c8bea149c276