puppetdb::app Add GC parameters to ro-host
It seems that the read only puppetdb servers still performs GC
operations and as they are missing theses settings they do the
operations using the default values.  As our overrides increase the
timeout it means that the parameters we specify are essentially ignored

This CR updates the config so the ro database  is also configured with
the correct TTL values

e.g. of GC on puppetdb2002 (trimed for gerrit)
  $ grep -vE 'store|replace' /var/log/puppetdb/puppetdb.log
  INFO  [p.p.c.services] Starting gc packages
  INFO  [p.p.c.services] Finished gc packages
  INFO  [p.p.c.services] Starting database garbage collection
  INFO  [p.p.c.services] Finished database garbage collection
  INFO  [p.p.c.services] Starting sweep of stale nodes (threshold: 7 days)
  INFO  [p.p.c.services] Finished sweep of stale nodes (threshold: 7 days)
  INFO  [p.p.c.services] Starting purge deactivated and expired nodes (threshold: 14 days)
  INFO  [p.p.c.services] Finished purge deactivated and expired nodes (threshold: 14 days)
  INFO  [p.p.c.services] Starting sweep of stale reports (threshold: 14 days)
  INFO  [p.p.c.services] Finished sweep of stale reports (threshold: 14 days)

Hosts: re:puppetdb[12]002.*wmnet
Change-Id: I3f4e44ebc96c3ed6b59626d65e0fcffd6c45e575