maintain-dbusers: rely on the global_id, not username for paws
In only four days, 7 users changed their wiki username, which confused
this service into thinking they were new accounts. That causes it to try
to overwrite an immutable file and fail (crash). Systemd is nice enough
to restart it, but that's hardly good. Since that doesn't actually mean
they deleted anything or whatever, this should really only use the UID
for PAWS, which is encoded in the mysql_username field.

This should prevent the problem from happening, though it would be more
elegant to use I514ed454a1939d7f2fb2bcec8 if the weird tool accounts
are fixed.

Bug: T276284
Change-Id: I47176c53251191f346485b703a3d785384bda366