hiera/nodes: put mc1037 into our session redis and memcached cluster
mc1037 will replace the dead mc1027 in shard09. This will not cause
a resharding in our memcached cluster but rather put this shard back
into production (gutter pool has replaced it so far). Regarding redis,
there will be the replacement of the current server in shard09, with
this one.

Puppet will be disabled while the host is running puppet, and then
we will let it take its course.

Bug: T278225
Change-Id: I9f2b9e0f32e75de9ff6afe6c4fb84ede3cc2fdb8