hiera/nodes: put mc1038 into our session redis and memcached cluster
mc1038 will replace the dead mc1024 in shard06. This will cause
a resharding in our memcached cluster since we are re-introducting
shard06 to mcrouter.  Regarding redis, there will be the replacement
of the current server in shard06.

Puppet will be disabled while the host is running puppet, and then
we will let it take its course.

Bug: T278225
Change-Id: Ib6a0c63f35a6b769462eeb67ac06327bfb28f47c