releases: Provide server-agnostic rsync for security patches
pipelinelib will invoke the rsync at build time to make sure we're
getting the latest patches. Since we want to avoid hardcoding the
name of a specific server, name the one intended for the primary
releases server "srv-patches-releases-primary".

When the primary server changes, puppet will adjust the target and
the blubber config won't need adjusting.

Change-Id: Id4907d014d0a902cf215407e901593b8a029ce55