gitlab: add gitlab.wm.org service IP, with lookup from Hiera
Previously gitlab.wikimedia.org was a CNAME for gitlab1001.wikimedia.org
but we dropped that in Ia75a707508c and then a new service IP was
assigned in netbox and then added to DNS via cookbook.

This is to add it to the actual interface on the VM gitlab1001.

Add proper parameters for v4 and v6 service IP to the profile class
and lookup values from Hiera. A little step towards doing it right
from the start without making it too complicated either.

Bug: T276148
Change-Id: I41f6900651bdadbd76537f59e30ae9c4aef9890c