gerrit: reported owner is actually patchset author
When a patchset is uploaded, the message reported in Phabricator has the
authorname which is mislabelled as being the change owner.  It leads to
confusion when someone send a patch authored by another person which
would be listed as the owner while the change is owned by whoever
uploaded the first patch.

Simply adjust the message from owner to author to match the field being
used.

Bug: T224262
Change-Id: Ia0170783a8d506bbf1855ef0058482767aaf804d