gerrit: escape remarkup for Phabricator comments [2]
Patchset events reported to Phabricator report the subject as is which
is then processed as remarkup. That leads to unwanted side effect such
as a variable named `__version__` being rendered as italic and with
underscores stripped.

Wrap the commit subject in a literal block markup (%%%). Do the same for
the reason stated when abandoning a change.

Bug: T93331
Change-Id: I346a3fb7480a21c67d3727b2c5bad1498e202dd7