gerrit: escape remarkup for Phabricator comments
Patchset events reported to Phabricator report the subject as is which
is then processed as remarkup. That leads to unwanted side effect such
as a variable named `__version__` being rendered as italic and with
underscores stripped.

Wrap the commit subject in a literal block markup (%%%). Do the same for
the reason stated when abandoning a change.

Bug: T93331
Change-Id: I2c3d866405485f7598f9db11adb8e1f72e1e0219