nova vendordata first boot script: avoid a race with puppet agents
With the new image build system, base images have puppet
already enabled and ready to run at boot time. If it's
a first boot, stop that puppet agent so we can run
it synchronously and wait on completion later.

Without this our attempt to force a puppet run during first
boot fails which causes us to enable logins before we're
completely ready.

Bug: T278051
Change-Id: I50474a0fe15c3b07ccc8eb0c89016a1e71d649d7