logstash: replace ECS allow list with filter_on_templates
71-filter_ecs_cleanup was generated by the ECS template build process
and necessitated a manual step to bring in the regenerated filter on
index template update.  With filter_on_templates, the allow list is
dynamically generated by the filter based on index templates.

filter_on_templates also enables tracking of dropped fields due to no
field definition or type mismatch by adding them to the normalized
object as an array of keywords attached to the event.  This satisfies
the requirement to track dropped fields as required in the Common
Logging Schema design document.

Bug: T234565
Change-Id: I166ca08de9d49002b6e5227f9872fe9fb81bd039