hadoop: increase the Yarn's defaults for vmem vs pmem ratio
In some Hadoop jobs the containers are killed if their virtual memory
usage is too high (compared to the 'pmem', namely the physical memory
allocated by the container). This change bumps the ratio to 10/1, to see
if there is any difference.

In https://issues.apache.org/jira/browse/YARN-4714 there is no clear
indication about what it is the gold standard, especially due to how
java/spark allocate virtual memory.

Bug: T278441
Change-Id: Icb3298ca64164e65348d2f4b4c16ba5945f4ed66