hadoop: introduce the GPU labels for the test cluster
The Yarn node labels are a good way to allow users to select Hadoop
workers. It needs the Capacity scheduler to work, and it can work in two
ways:
1) all the node managers communicate to the RM their labels (the
   decentralized way).
2) the RM is instructed with what nodes have label x, via command line
   (didn't find a way to store a static list) and it stores it on HDFS
   (convenient so the state is in one place for both RMs).
This change introduces the latter, together with a minimal config for
the capacity scheduler, to address the GPUs use case (allow researchers
to run jobs only on nodes equipped with a GPU). In the future we may
want something more elaborate, but I'd keep it simple for the moment.

Bug: T264401
Change-Id: If969892855263b52689651840d783cc353200fbf