C:ssh::server: update parameter types
This CR changes the type of the port parameter and moves simplifies the
default logic for `$authorized_keys_file`.  This CR is a noop in
production however pc will show a change due to the difference in the
parameters.

I have checked the instance-puppet repo and cloud should not be affected
by this cr

Hosts: cumin1001.eqiad.wmnet
Change-Id: Ia6a0d2e5fe5282d15e6bbf63d4532b498f76a3e2