openstack: l3_agent: conntrackd: stop using systemd Watchdog
The conntrackd daemon is very busy dealing with neutron l3 agent, and it fails
to report to systemd that its alive, even though it is. Systemd kills the
daemon and restarts it, which doesn't help.

Remove the Watchdog option to try leaving room for the daemon for doing
its job.

Bug: T268335
Change-Id: I905e56a73b0cff6c6e32eaf12c0fcc59dbc783dd
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>