openstack: neutron: reduce net.netfilter.nf_conntrack_tcp_timeout_established
The default net.netfilter.nf_conntrack_tcp_timeout_established value is 5 days,
which is too long of a value. With 1 day it should be enough to properly cover
our NAT use case.

In our current HA setup, with conntrackd, having conntrack entries with such a
long timeout leaves the replica router with a lot of bogus entries. This should
make the nf_conntrack in-kernel garbage collector kick-in earlier.

Bug: T268335
Change-Id: I5e7b8448bf70f55283f8fcfd80f3b52281d6aca5
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>