contint: serve compressed json as application/json
Fresnel generates traces in json traces which we want to compress on the
server side. We also need to be able to drag'n drop them in Chrome
console which requires application/json type.

This results in requests to the trace json.gz file to be emitted
with the headers:

    content-type: application/json
    content-encoding: gzip

Which let us drag'n drop a Fresnel performance report to Chrome
performance tab.

Bug: T249268
Change-Id: Ibdd489573e178e7a95e07c0c09e70ed8cec44c8c