mailman3: Add documentation to classes, merge hyperkitty into web
hyperkitty is installed as part of mailman-web, so there's no need to
install its package separately. It also needs the rest of the web
component, so there's no real value in having it as a separate class.

My main motiviation for this is that it will make future refactoring
simpler.

Change-Id: I8432270790ac454e8ebe114d3c9d97503e4d5eb6