hive: use the default logging config path for parquet logs
We think our custom config file is not being loaded due to the way
HADOOP_CLIENT_OPTS is constructed, and that
java.util.logging.config.file will be set to point to
conf/parquet-logging.properties no matter how we tweak the
environment variable. Instead of trying to override the config file
location, this patch moves our config to the expected path.

Bug: T275757
Change-Id: I255421b76335af3b247b86da03058d93a60eb341