kubectl: Fetch it from_future for a set of hosts
Why:
Now that production is updated, we want to upgrade also the clients that
talk to it

What:
Split off kubernetes-client from master and helm modules to a specific
class of the k8s/ modules
Create a specific profile class that is included where needed alongside
the necessary hiera keys. Also update the deployment hosts respectively

The CI hosts will get it via the profile::kubernetes::deployment_server
host, but releases is on purpose not being changed here. The package
will be manually purge from those hosts

Bug: T278356
Change-Id: Ifaf235e49717bcdef29faef3d2a81e490115a73d