ssh::server us legacy facts for ip addresses
Currently facter is not that intelligent at picking what it considers to
be the primary ip addresses.  As such we have overridden the legacy
ipaddress and ipaddress6 facts[1] so that they point to the correct
addresses.

We are not able to override the networking structured fact
in the same way we override the legacy fact.  As such for now revert to
using the legacy facts

The following upstream bug is related to this and a possible way forward
to fixing this issue
 https://tickets.puppetlabs.com/browse/FACT-2907

[1]modules/base/lib/facter/interface_primary.rb

Hosts: contint1001.wikimedia.org,gerrit1001.wikimedia.org
Change-Id: If7b28b7b223a4e5a53e40946826fcc311bf7fa85