wmfkeystonehooks ldap groups: Handle groups with no members
Previously if we created a group with no members or removed
the last member from a group, an exception was raised.

This never happened prior to

  https://gerrit.wikimedia.org/r/c/operations/puppet/+/667423

...but now it happens quite a lot.

Since a group can't exist in ldap without members, we just postpone
group creation until the first member is added, and delete the group
if/when the last member is removed.

Bug: T279491
Change-Id: Ia6cb9440375efbf23b24573f2533004ff71035b8