service::deploy::gitclone: do not use mediawiki/services as prefix
When deploying a service::node service without scap (for example, aqs in
the Analytics wmcs namespace) the gerrit repository can be anything, not
necessarily prefixed with "mediawiki/services".
The service::deploy::gitclone define seems not used in production, and
probably it is not even in wmcs since most of the service::node instance
are using scap (that Analytics cannot use in its wmcs namespace since it
would require extra vms for this single use case).

Change-Id: Ie4cbb4023d8ea7f2aba357dbfa87e48b2a19828d