hadoop: add the liblog4j-extras1.2-java jar to HADOOP_CLASSPATH
This change adds apache-log4j-extras.jar to the HADOOP_CLASSPATH on
every node using it (the package is available for Debian 9/10/11). We
can selectively enable it only where daemons are running, to avoid
deploying it on client nodes, but it may be a little subset of the whole
infrastructure. In theory with this change we'll be able to use the
RollingFileAppender in log4j configs everywhere, to automatically gzip
log files etc..

Bug: T276906
Change-Id: I407a8dd9872eb1abbe74d8eed6b383ffbc44eef5